pyproject.toml
[project]
name = "auto_k8s_pilot"
version = "0.1.0"
description = "auto-k8s-pilot using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.165.1,<1.0.0",
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # OpenTelemetry
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    # Logging
    "structlog>=24.4.0",
    # Utils
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
auto_k8s_pilot = "auto_k8s_pilot.main:run"
run_crew = "auto_k8s_pilot.main:run"
api = "auto_k8s_pilot.api:main"
train = "auto_k8s_pilot.main:train"
replay = "auto_k8s_pilot.main:replay"
test = "auto_k8s_pilot.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_k8s_pilot"]

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "types-pyyaml>=6.0.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
